Extending the publication of mu-cal-scope-ids
To have a more fine grained control in a mu-stack over what (internal)
deltas needs to go to what services, the call scopes have been extended.

If properly configured in the rules of the (internal) deltanotifier,
this can reduce the load on the system.

This allows also to simplify the code in charge of the problematic
ignoreFromSelf behaviour in certain deploys of the stack.

Furthermore, this allows
better coordination when multiple producers with overlapping config
work on the system.

E.g.
Previously, when a person is added to one publication graph, this
fired an event which could be consumed by another producer.

In some specific sequence of notifications, this led to an infinite interaction
of producers, leading to infinte deltas. (Eventually cut off by a
healing process, but that is mere coincidence)

For documentation, the case was.
 - producer 1 (P1)
 - producer 2 (P2)
 - Statement 1 (S1): "ADD <foo> <bar> <baz>"
 - Statement 2 (S2): "DELETE <foo> <bar> <baz>"

At some point the following happend:

P1 exec S1 at t1, P2 exec S2 at t2

this led to

P1 exec S2 at t3, P2 exec S1 at t4

etc...

Note: we'll need even more fine grained filtering of the deltas to
deduce what is really relevant for the producer. This will be for a
next commit.