Slight change of approach for DELTA_CHUNK_SIZE
We aim for a more granular method to throttle the chunk size.
For some predicates of specific types, the provided DELTA_CHUNK_SIZE
is too large, which continues to cause memory explosions in node.js

For instance, the case 'dct:description' for type 'besluit:Besluit'
currently returns only 800k triples.
However, the strings can be exceedingly large, resulting in arrays about 500 MB in size.

Consequently, we've introduced the ability to throttle queryChunkSize
per predicate and per type.
The implementation might look like this:
(note that the naming might evolve over commits, but the idea will remain the same):

```
{
  "export": [
    { "type": "http://data.vlaanderen.be/ns/besluit#Artikel",
      "graphsFilter": [ "http://mu.semte.ch/graphs/public" ],
      "properties": [ "http://data.europa.eu/eli/ontology#title" ],
      "healingOptions": {
        "http://data.europa.eu/eli/ontology#title": {
          "queryChunkSize": 55555
        }
      }
    }
  ]
}
```