WIP: An attempt at better syncing publication files <--> graph
This is a failed attempt at better syncing the publication graph and files, by first trying to publish the files and then updating the publication graph.

However, this fails in the following scenario: Suppose the published files ended up with this statement:

```
DELETE triple_1
```
In a second step, somehow the publication graph update failed and doesn't include the last statement. You would expect healing to take care of this.

However, the file is already ready for consumption and was consumed by a consumer before the next healing run.

In the meantime, before the healing, in the source, the following statement was made:

```
INSERT triple_1
```

Since triple_1 is in the publication graph, the files won't contain the equivalent of INSERT triple_1 (since it's still in the publication graph). This means the files (or rather the full replay of the inserts/deletes in the files) will result in:

```
DELETE triple_1
```

Whereas, based on the source, it should be:

```
INSERT triple_1
```

So, basically, it's still broken.

The only thing I could come up with is that the publication of the files AND the updating of the publication graph should happen in one go (so basically a transaction). But that's a bit tricky since the triples of the publication graph (might) reside in another store. Furthermore, even if in the same store, I am not sure mu-auth will perform it as a transactional statement.

So the alternative is a very complicated rollback/n-step-publication procedure, but this might get very complicated.